% Generated by roxygen2 (4.0.1): do not edit by hand
\name{c_tabulate}
\alias{c_tabulate}
\title{Function to take a vector of species IDs and species names, and calculate the number of each species
occuring in the vector. For use with \code{\link{stoch_patch}} model output.}
\usage{
c_tabulate(spec_vec, spec_names)
}
\arguments{
\item{spec_vec}{An integer vector of species IDs}

\item{spec_names}{A character vector of species names to which the \code{spec_vec} acts as an index}
}
\value{
A named vector, tabulating the number of times each species occurs in the ID vector.
}
\description{
Function to take a vector of species IDs and species names, and calculate the number of each species
occuring in the vector. For use with \code{\link{stoch_patch}} model output.
}

