% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gen_CR_matrix}
\alias{gen_CR_matrix}
\title{Function to generate consumer-resource consumption probability matrix with phylogenetic structure}
\usage{
gen_CR_matrix(ctree = rcoal(15), rtree = rcoal(25), sig = 1,
  phi = c(0.2, 0.2, 0, 2, 0.2, 0.2), alpha = 0)
}
\arguments{
\item{ctree}{A phylogenetic tree of class phylo, representing the phylogenetic relationships
for the consumers}

\item{rtree}{A phylogenetic tree of class phylo, representing the phylogenetic relationships
for the resources}

\item{sig}{Total Variance to be generated in the latent variables via the phylogenies}

\item{phi}{Unit simplex vector (elements between 0 and 1 and sum to 1) describing how the
phylogenetic variance is partitioned into different kinds of phylogenetic variability.
The elements correspond to:
\itemize{
 \item{1) Phylogenetic structure in resource richness}
 \item{2) Phylogenetic structure in consumer range}
 \item{3) Phylogenetic structure in resources' consumer sets}
 \item{4) Phylogenetic structure in consumers' resource sets}
 \item{5) Phylogenetic interactive structure between consumer and resource phylogenies}
}}

\item{alpha}{Intercept for latenet variables (higher values mean probabilities are uniformly
larger).}
}
\value{
A consumer-resource consumption probability matrix with consumers in the rows, and
resources in the columns
}
\description{
Function to generate consumer-resource consumption probability matrix with phylogenetic structure
}

