% Generated by roxygen2 (4.0.0): do not edit by hand
\name{make_OD_comm}
\alias{make_OD_comm}
\title{Function to generate a phylogenetically overdispersed community}
\usage{
make_OD_comm(pdist, n.spp, prop)
}
\arguments{
\item{pdist}{Phylogenetic distance matrix (as generated by \code{\link{cophenetic}})
Row names must contain the species names.}

\item{n.spp}{Number of species to randomly sample from full species pool}

\item{prop}{Proportion of randomly sampled species to keep at end of overdispersion
algorithm}
}
\description{
Function to generate a phylogenetically overdispersed community
}
\details{
This function iteratively removed one species from the pair with minimum distance
until the number of remaining species equals \code{n.spp*prop}. The initial random sampling
makes sure there is some stochasticity to the resulting communities.
}

