% Generated by roxygen2 (4.0.0): do not edit by hand
\name{make_OD_comms}
\alias{make_OD_comms}
\title{Function to generate a phylogenetically overdispersed communities}
\usage{
make_OD_comms(pdist, n.spp, prop)
}
\arguments{
\item{pdist}{Phylogenetic distance matrix (as generated by \code{\link{cophenetic}}).
Row names must contain the species names.}

\item{n.spp}{Integer vector specifying the umber of species to randomly sample from
full species pool, for each community.}

\item{prop}{Numeric vector specifying proportion of randomly sampled species to keep
at end of overdispersion algorithm, for each community.}
}
\description{
Function to generate a phylogenetically overdispersed communities
}
\details{
This function generates multiple overdispersed communities by iteratively
removed one species from the pair with minimum distance until the number of remaining
species equals \code{n.spp*prop}, where \code{n.spp} and \code{prop} are vectors
of length equal to the number of desired communities. The initial random sampling
makes sure there is some stochasticity to the resulting communities.
}

