% Generated by roxygen2 (4.0.0): do not edit by hand
\name{make_rand_comms}
\alias{make_rand_comms}
\title{Function to generate a phylogenetically random communities.}
\usage{
make_rand_comms(pdist, n.spp, prop)
}
\arguments{
\item{pdist}{Phylogenetic distance matrix (as generated by \code{\link{cophenetic}}).
Row names must contain the species names.}

\item{n.spp}{Integer vector specifying the number of species to randomly sample from
full species pool, for each community.}

\item{prop}{Numeric vector specifying proportion of randomly sampled species to keep
for each community.}
}
\description{
Function to generate a phylogenetically random communities.
}
\details{
This function generates multiple random communities by  randomly
choosing \code{n.spp*prop} species.
}

