% Generated by roxygen2 (4.0.1): do not edit by hand
\name{make_UD_comm}
\alias{make_UD_comm}
\title{Function to generate a phylogenetically underdispersed community}
\usage{
make_UD_comm(pdist, n.spp, prop)
}
\arguments{
\item{pdist}{Phylogenetic distance matrix (as generated by \code{\link{cophenetic}})
Row names must contain the species names.}

\item{n.spp}{Number of species to randomly sample from full species pool}

\item{prop}{Proportion of randomly sampled species to keep at end of underdispersion
algorithm}
}
\description{
Function to generate a phylogenetically underdispersed community
}
\details{
This function starts with a randomly chosen species, and then adds the most
closely related species. It then adds the most closely related species to the newly
chosen species, iterating through this procedure until the number of added species equals
\code{n.spp*prop}. The initial random sampling adds extra stochasticity to the resulting
communities.
}

