% Generated by roxygen2 (4.0.1): do not edit by hand
\name{make_UD_comms}
\alias{make_UD_comms}
\title{Function to generate a phylogenetically underdispersed communities}
\usage{
make_UD_comms(pdist, n.spp, prop)
}
\arguments{
\item{pdist}{Phylogenetic distance matrix (as generated by \code{\link{cophenetic}}).
Row names must contain the species names.}

\item{n.spp}{Integer vector specifying the number of species to randomly sample from
full species pool, for each community.}

\item{prop}{Numeric vector specifying proportion of randomly sampled species to keep
at end of overdispersion algorithm, for each community.}
}
\description{
Function to generate a phylogenetically underdispersed communities
}
\details{
This function generates multiple underdispersed communities by starting
with a randomly chosen species, and then adds the most closely related species.
It then adds the most closely related species to the newly chosen species, iterating
through this procedure until the number of added species equals \code{n.spp*prop}.
The initial random sampling adds extra stochasticity to the resulting communities.
}

