% Generated by roxygen2 (4.0.0): do not edit by hand
\name{plot_CR_matrix}
\alias{plot_CR_matrix}
\title{Function to plot a consumer-resource consumption probability matrix, along with
phylogenies}
\usage{
plot_CR_matrix(ctree, rtree, prob_mat)
}
\arguments{
\item{ctree}{A phylogenetic tree of class phylo, representing the phylogenetic relationships
for the consumers}

\item{rtree}{A phylogenetic tree of class phylo, representing the phylogenetic relationships
for the resources}

\item{prob_mat}{Consumer-resource consumption probability matrix like that produced
by \code{\link{gen_CR_matrix}}.}
}
\description{
Function to plot a consumer-resource consumption probability matrix, along with
phylogenies
}

