% Generated by roxygen2 (4.0.0): do not edit by hand
\name{stoch_patch}
\alias{stoch_patch}
\title{Stochastic Patch model with patch "types"}
\usage{
stoch_patch(t, y, params)
}
\arguments{
\item{t}{Timestep. Not used in the simulation model, but required for using \code{\link{deSolve}}}

\item{y}{Current state of the model. This is a vector with length equal to the number of patches in the site.
Each element is an integer that indexes the "species" that occupies that patch. 0 represents an empty patch.}

\item{params}{A list of parameter for the simulation. The parameters are as follows:
\itemize{
 \item{"d" - The death rate. Rate at which occupied patches become unoccupied}
 \item{"nspec" - The number of "species" in the total species pool}
 \item{"poolprob" - The probability that an empty patch will be colonized by a species from the pool}
 \item{"ntype" - The number of different patch "types"}
 \item{"occmat" - A \code{nspec} by \code{ntype} matrix containing the occupation probability of each
 species on each patch type. Generally, each element is either a 0 or a 1, where 0 means the species cannot
 occupy that type and 1 means that it can.}
 \item{"type" - Patch types. A vector of integer with length equal to the number of patches. Each element
 is an index to the patch types}
}}
}
\value{
A vector of integers indexing the species occupying each patch after the model runs for one timestep
}
\description{
Stochastic Patch model with patch "types"
}

